Rename logExpander to logItem
