+++ src/components/logdisplay/LogItem.vue
